Commit test progress
